+++ src/engine/layouts/kwin.ts
@@ -1,7 +1,13 @@
 // kwin.ts - Layout engine that mimics kwin default tiling
 // quite a bit smaller this time!
 
-import { TilingEngine, EngineCapability, Client, Tile, EngineSettings } from "../engine";
+import {
+    TilingEngine,
+    EngineCapability,
+    Client,
+    Tile,
+    EngineSettings,
+} from "../engine";
 import { Direction } from "../../util/geometry";
 
 export default class KwinEngine extends TilingEngine {
@@ -11,7 +17,7 @@ export default class KwinEngine extends TilingEngine {
     get engineSettings(): EngineSettings {
         return {};
     }
-    set engineSettings(s: EngineSettings) {}
+    set engineSettings(_: EngineSettings) {}
 
     buildLayout() {
         // literally nothing