+++ src/engine/layouts/monocle.ts
@@ -1,6 +1,12 @@
 // monocle.ts - Monocle layout
 
-import { TilingEngine, EngineCapability, Client, Tile, EngineSettings } from "../engine";
+import {
+    TilingEngine,
+    EngineCapability,
+    Client,
+    Tile,
+    EngineSettings,
+} from "../engine";
 import { Direction } from "../../util/geometry";
 
 export default class MonocleEngine extends TilingEngine {
@@ -8,8 +14,8 @@ export default class MonocleEngine extends TilingEngine {
     clients: Client[] = [];
     get engineSettings(): EngineSettings {
         return {};
-    };
-    set engineSettings(s: EngineSettings) {}
+    }
+    set engineSettings(_: EngineSettings) {}
 
     buildLayout() {
         this.rootTile = new Tile();
@@ -33,7 +39,7 @@ export default class MonocleEngine extends TilingEngine {
     }
 
     // handle switching order of windows through side-based insertion
-    // inserting above/right puts window on top, inserting 
+    // inserting above/right puts window on top, inserting
     putClientInTile(client: Client, _tile: Tile, direction?: Direction) {
         if (this.clients.includes(client)) {
             return;
@@ -44,8 +50,8 @@ export default class MonocleEngine extends TilingEngine {
         }
         // if up or right then cycle forwards (assuming the window inserted was previously on top)
         if (
-            ((direction & Direction.Up) && (direction & Direction.Vertical)) ||
-            ((direction & Direction.Right) && !(direction & Direction.Vertical))
+            (direction & Direction.Up && direction & Direction.Vertical) ||
+            (direction & Direction.Right && !(direction & Direction.Vertical))
         ) {
             // turns out this just means inserting the window normally
             // (its automatically pushed to the back)